Added test for cleaning directoriez closes #97

+++ Sources/Core/PackageCleaner.cs
@@ -15,9 +15,16 @@
         {
             MachineCache.Default.Clear();
             var dir = Path.Combine(Path.GetTempPath(), "nuget");
-            Directory.Delete(dir,true);
+            if (Directory.Exists(dir))
+            {
+                Directory.Delete(dir, true);
+            }
+
             var dir2 = Path.Combine(Path.GetTempPath(), "NuGetScratch");
-            Directory.Delete(dir2, true);
+            if (Directory.Exists(dir2))
+            {
+                Directory.Delete(dir2, true);
+            }
         }
     }
 }
